型統合リファクタリング: 共有値型パターンによるネスト型の統合
CharacterSnapshotとRuntimeCharacterProgressの重複ネスト型を
CharacterValues enumに集約し、RuntimePartyProgressを削除。

- CharacterValues.swift新規作成（9つのネスト型を定義）
- CharacterSnapshot/RuntimeCharacterProgress: typealiasに変更
- PartySnapshot: memberIds computed property追加
- RuntimePartyState: PartySnapshotを直接使用
- ProgressRuntimeService: 変換コード簡素化（約50行削減）
