+++ Epika/Progress/Domain/Snapshots/PartySnapshot.swift
@@ -10,4 +10,7 @@ struct PartySnapshot: Sendable, Hashable {
     var targetFloor: UInt8
     var memberCharacterIds: [UInt8]            // 順序=配列index
     var updatedAt: Date
+
+    /// RuntimePartyProgress互換のプロパティ名
+    var memberIds: [UInt8] { memberCharacterIds }
 }
