+++ Epika/Services/GameRuntime/Party/RuntimePartyState.swift
@@ -1,14 +1,5 @@
 import Foundation
 
-struct RuntimePartyProgress: Sendable, Hashable {
-    var id: UInt8                              // 1〜8
-    var displayName: String
-    var lastSelectedDungeonIndex: UInt16       // 0=未選択
-    var lastSelectedDifficulty: UInt8
-    var targetFloor: UInt8
-    var memberIds: [UInt8]                     // 順序=配列index
-}
-
 struct RuntimePartyState: Sendable {
     struct Member: Identifiable, Sendable {
         var id: UInt8 { characterId }
@@ -17,10 +8,10 @@ struct RuntimePartyState: Sendable {
         let character: RuntimeCharacterState
     }
 
-    let party: RuntimePartyProgress
+    let party: PartySnapshot
     let members: [Member]
 
-    init(party: RuntimePartyProgress, characters: [RuntimeCharacterState]) throws {
+    init(party: PartySnapshot, characters: [RuntimeCharacterState]) throws {
         self.party = party
         let characterMap = Dictionary(uniqueKeysWithValues: characters.map { ($0.progress.id, $0) })
         var mappedMembers: [Member] = []
